src/requirements.txt
albumentations==1.1.0
click==8.0.3
imgaug==0.4.0
joblib==1.1.0
matplotlib==3.5.1
multiprocess==0.70.14
numpy==1.22.3
opencv-python==4.5.5.62
opencv-python-headless==4.5.5.62
pandas==1.3.5
Pillow==8.4.0
progress==1.6
PyYAML==6.0
scikit-image==0.19.1
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
timm==0.6.7
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.0
